Refactor: Moved methods from ItemChecker to ItemNamingUtils.
