+++ src/main/java/com/Acrobot/ChestShop/Commands/ItemInfo.java
@@ -8,8 +8,8 @@ import com.Acrobot.Breeze.Utils.StringUtil;
 import com.Acrobot.ChestShop.ChestShop;
 import com.Acrobot.ChestShop.Configuration.Messages;
 import com.Acrobot.ChestShop.Events.ItemInfoEvent;
-import com.Acrobot.ChestShop.Listeners.PreShopCreation.ItemChecker;
 import com.Acrobot.ChestShop.Utils.ComponentUtils;
+import com.Acrobot.ChestShop.Utils.ItemNamingUtils;
 import net.md_5.bungee.api.chat.TextComponent;
 import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
@@ -55,8 +55,8 @@ public class ItemInfo implements CommandExecutor {
         tc.setColor(net.md_5.bungee.api.ChatColor.WHITE);
         tc.addExtra(ComponentUtils.getLocalizedItemName(item));
 
-        String fullName = ChatColor.stripColor(ItemChecker.getDisplayName(item));
-        String signItemName = ChatColor.stripColor(ItemChecker.getSignItemName(item));
+        String fullName = ChatColor.stripColor(ItemNamingUtils.getDisplayName(item));
+        String signItemName = ChatColor.stripColor(ItemNamingUtils.getSignItemName(item));
         TextComponent signNames = new TextComponent("  ");
         signNames.setColor(net.md_5.bungee.api.ChatColor.DARK_GRAY);
         signNames.addExtra(fullName);
