+++ src/test/java/com/Acrobot/ChestShop/Tests/ItemNamingUtilsTest.java
@@ -1,16 +1,17 @@
 package com.Acrobot.ChestShop.Tests;
 
-import com.Acrobot.ChestShop.Listeners.PreShopCreation.ItemChecker;
 import static org.junit.Assert.assertEquals;
+
+import com.Acrobot.ChestShop.Utils.ItemNamingUtils;
 import org.junit.Test;
 
-public class ItemCheckerTest {
+public class ItemNamingUtilsTest {
 
     @Test
     public void testNamingCutdown() {
         String sut = "Hello there! Didn't your tests fail?";
 
-        String result = ItemChecker.shortenDisplayName(sut, 15);
+        String result = ItemNamingUtils.shortenDisplayName(sut, 15);
 
         assertEquals("HelTheDidYoTeFa", result);
     }
@@ -19,7 +20,7 @@ public class ItemCheckerTest {
     public void testNamingCutdownWhenSingleWordsAreShorter() {
         String sut = "Hi i am your, testededed!";
 
-        String result = ItemChecker.shortenDisplayName(sut, 15);
+        String result = ItemNamingUtils.shortenDisplayName(sut, 15);
 
         assertEquals("HiIAmYour,Teste", result);
     }
@@ -28,7 +29,7 @@ public class ItemCheckerTest {
     public void testNamingCutdownOnShortEnough() {
         String sut = "Hi i am tested";
 
-        String result = ItemChecker.shortenDisplayName(sut, 15);
+        String result = ItemNamingUtils.shortenDisplayName(sut, 15);
 
         assertEquals(sut, result);
     }
@@ -37,7 +38,7 @@ public class ItemCheckerTest {
     public void testNamingOnlyRemoveWhitespaces() {
         String sut = "Hi i am all tested";
 
-        String result = ItemChecker.shortenDisplayName(sut, 15);
+        String result = ItemNamingUtils.shortenDisplayName(sut, 15);
 
         assertEquals("HiIAmAllTested", result);
     }
